requirements.txt
mssql-django>=1.5
pyodbc>=5.1
